moved gmm_eigen.h to shared
