+++ src/cpp/shared/gmm_eigen.h
